get location at MainActivity
