+++ app/src/main/java/com/wafflestudio/shafe/MapActivity.java
@@ -11,26 +11,16 @@ import android.util.Log;
 import android.view.ViewGroup;
 import android.widget.Toast;
 
-import com.google.android.gms.common.ConnectionResult;
-import com.google.android.gms.common.api.GoogleApiClient;
-import com.google.android.gms.common.api.GoogleApiClient.ConnectionCallbacks;
-import com.google.android.gms.common.api.GoogleApiClient.OnConnectionFailedListener;
-import com.google.android.gms.location.LocationServices;
 import android.Manifest;
 
 import butterknife.Bind;
 import butterknife.ButterKnife;
 
+public class MapActivity extends AppCompatActivity {
 
-public class MapActivity extends AppCompatActivity implements
-        ConnectionCallbacks, OnConnectionFailedListener {
+    protected static final String TAG = "MapActivity";
 
-    protected static final String TAG = "MainActivity";
-
-    protected GoogleApiClient mGoogleApiClient;
-    protected Location mLastLocation;
-
-    protected MapManager searchResultMap;
+    protected MapManager searchResultMap = MapManager.getInstance();
 
     private static final int REQUEST_CODE_LOCATION = 2;
 
@@ -51,12 +41,6 @@ public class MapActivity extends AppCompatActivity implements
 
         //enable up button
         getSupportActionBar().setDisplayHomeAsUpEnabled(true);
-
-        buildGoogleApiClient();
-
-        searchResultMap = MapManager.getInstance(this);
-
-        mapViewContainer.addView(searchResultMap.getMapView());
     }
 
     public void onRequestPermissionsResult(int requestCode, String[] permissions, int[] grantResults) {
@@ -70,72 +54,21 @@ public class MapActivity extends AppCompatActivity implements
         }
     }
 
-    /**
-     * Builds a GoogleApiClient. Uses the addApi() method to request the LocationServices API.
-     */
-    protected synchronized void buildGoogleApiClient() {
-        mGoogleApiClient = new GoogleApiClient.Builder(this)
-                .addConnectionCallbacks(this)
-                .addOnConnectionFailedListener(this)
-                .addApi(LocationServices.API)
-                .build();
-    }
-
     @Override
     protected void onStart() {
         super.onStart();
-        mGoogleApiClient.connect();
+
+        searchResultMap.createMapView(this);
+        searchResultMap.initializeMapWithSavedValue();
+
+        mapViewContainer.addView(searchResultMap.getMapView());
     }
 
     @Override
     protected void onStop() {
         super.onStop();
-        if (mGoogleApiClient.isConnected()) {
-            mGoogleApiClient.disconnect();
-        }
+
         // TODO : save를 어느 시점에서 할지 고민 해볼 것
         searchResultMap.saveMapCenterPointAndZoomLevel();
-    }
-
-    /**
-     * Runs when a GoogleApiClient object successfully connects.
-     */
-    @Override
-    public void onConnected(Bundle connectionHint) {
-        // Provides a simple way of getting a device's location and is well suited for
-        // applications that do not require a fine-grained location and that do not need location
-        // updates. Gets the best and most recent location currently available, which may be null
-        // in rare cases when a location is not available.
-
-        if (ActivityCompat.checkSelfPermission(this, android.Manifest.permission.ACCESS_FINE_LOCATION)
-                != PackageManager.PERMISSION_GRANTED) {
-            ActivityCompat.requestPermissions(this, new String[]{Manifest.permission.ACCESS_FINE_LOCATION}, REQUEST_CODE_LOCATION);
-        }
-        mLastLocation = LocationServices.FusedLocationApi.getLastLocation(mGoogleApiClient);
-
-        if (mLastLocation != null) {
-            searchResultMap.loadCurrentLocation(mLastLocation.getLatitude(), mLastLocation.getLongitude(), false);
-        }
-        else {
-            searchResultMap.loadCurrentLocation(37.450254, 126.952557, false);
-            Toast.makeText(this, R.string.no_location_detected, Toast.LENGTH_LONG).show();
-        }
-        searchResultMap.initializeMapWithSavedValue();
-    }
-
-    @Override
-    public void onConnectionFailed(ConnectionResult result) {
-        // Refer to the javadoc for ConnectionResult to see what error codes might be returned in
-        // onConnectionFailed.
-        Log.i(TAG, "Connection failed: ConnectionResult.getErrorCode() = " + result.getErrorCode());
-    }
-
-
-    @Override
-    public void onConnectionSuspended(int cause) {
-        // The connection to Google Play services was lost for some reason. We call connect() to
-        // attempt to re-establish the connection.
-        Log.i(TAG, "Connection suspended");
-        mGoogleApiClient.connect();
     }
 }
